+++ client/modules/Facility/FacilityViewDetail.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDom from "react-dom";
 import {ReactMeteorData} from 'meteor/react-meteor-data';
-import '../Request/PMPList.jsx';
+import '../Compliance/PMPList.jsx';
 
 ConfigureAdHoc = React.createClass({
     render() {